+++ packages/international-phone-input/src/types.ts
@@ -101,7 +101,7 @@ export type InternationalPhoneInputDesktopProps = CommonPhoneInputProps &
           >)
         | ({ options: InputAutocompleteDesktopProps['options'] } & Omit<
               InputAutocompleteDesktopProps,
-              'onChange' | 'onInput'
+              'onChange' | 'onInput' | 'onClear'
           >)
     );
 
@@ -113,7 +113,7 @@ export type InternationalPhoneInputMobileProps = CommonPhoneInputProps &
           >)
         | ({ options: InputAutocompleteMobileProps['options'] } & Omit<
               InputAutocompleteMobileProps,
-              'onChange' | 'onInput'
+              'onChange' | 'onInput' | 'onClear'
           >)
     );