+++ packages/bank-card/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Change Log
 
+## 5.8.21
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.39
+
+## 5.8.20
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.38
+
+## 5.8.19
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.37
+
+## 5.8.18
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   masked-input@6.3.36
+
 ## 5.8.17
 
 ### Patch Changes