+++ packages/custom-button/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 3.12.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.11.10
+
+## 3.12.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.11.9
+
+## 3.12.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   button@11.11.8
+
 ## 3.12.6
 
 ### Patch Changes