+++ packages/notification-manager/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 5.8.13
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   portal@3.3.10
+    -   notification@7.5.13
+
+## 5.8.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   portal@3.3.9
+    -   notification@7.5.12
+
+## 5.8.11
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   portal@3.3.8
+    -   notification@7.5.11
+
 ## 5.8.10
 
 ### Patch Changes