+++ packages/scrollbar/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 3.4.0
+
+### Minor Changes
+
+<sup><time>23.06.2025</time></sup>
+
+### [#1747](https://github.com/core-ds/core-components/pull/1747)
+
+-   Добавлен дополнительный класс для контейнера горизонтальной полосы прокрутки horizontalBarClassName
+
+## 3.3.0
+
+### Minor Changes
+
+<sup><time>23.05.2025</time></sup>
+
+### [#1679](https://github.com/core-ds/core-components/pull/1679)
+
+-   Добавлен ref для вертикальной полосы прокрутки
+
 ## 3.2.0
 
 ### Minor Changes