+++ packages/slider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 4.8.0
+
+### Minor Changes
+
+<sup><time>16.05.2025</time></sup>
+
+### [#1684](https://github.com/core-ds/core-components/pull/1684)
+
+-   Добавлен props snap для точного шага слайдера по range
+
 ## 4.7.3
 
 ### Patch Changes