+++ packages/spinner/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 4.0.9
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+
+## 4.0.8
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.1
+
+## 4.0.7
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.0
+
 ## 4.0.6
 
 ### Patch Changes